requirements.txt
python-telegram-bot[job-queue]
google-generativeai
python-dotenv
aiohttp
feedparser
beautifulsoup4
